REFACTOR :: 원서 접수 페이지에 전역 상태, stores/form/ 폴더로 이동
