+++ apps/user/src/app/form/성적입력/성적입력.tsx
@@ -12,13 +12,13 @@ import { UnderLineButton } from '@maru/ui';
 import { flex } from '@maru/utils';
 import { useState } from 'react';
 import { styled } from 'styled-components';
-import { useCTAButton } from './성적입력.hooks';
 import {
     useAttendanceInfoState,
     useCertificateListInfoState,
     useStudentSubjectListState,
     useVolunteerInfoState,
-} from './성적입력.state';
+} from '../../../stores/form/성적입력.state';
+import { useCTAButton } from './성적입력.hooks';
 
 const FIELD_DATA = ['성적 입력', '출결상황', '봉사시간', '자격증'] as const;
