+++ apps/user/src/app/form/전형선택/전형선택.tsx
@@ -3,8 +3,8 @@ import { FormLayout } from '@/layouts';
 import { RadioGroup } from '@maru/ui';
 import { flex } from '@maru/utils';
 import { styled } from 'styled-components';
+import { useChoiceFormTypeState } from '../../../stores/form/전형선택.state';
 import { useCTAButton, useInput } from './전형선택.hooks';
-import { useChoiceFormTypeState } from './전형선택.state';
 
 const 전형선택 = () => {
     const { choiceFormType } = useChoiceFormTypeState();
