+++ apps/user/src/app/form/초안작성완료/초안작성완료.tsx
@@ -195,6 +195,7 @@ const CompleteAlarmBox = styled.div`
         to {
             transform: translateY(-300px);
             opacity: 0;
+            display: none;
         }
     }
 `;
